Google ADK Hackathon Initial commit
I have entered into a Hackathon with Google!

This Hackathon is meant to showcase the ability of Multi-agent systems
in solving problems and discovering new solutions with these innovative
little pieces of technology.

I am going to be using Google Agent Development Kit (ADK)
to create a multi-agent system.

For now, this is just uploading a tutorial,
and the next commit will also be loading a tutorial.

In this commit, I have ommitted the file to pip install,
but the next commit will include all of them,
since the full tutorial should show me everything that
might need to be installed