Proof of Concept that we probably don't need chime
The team was considering using Chime for the multi_tool_agent,
but I wanted to make sure that we coudl use the existing agents
to analyze audio and video files and understand the
tone, content, and facial expressions of the people in the video.
I created a proof of concept that uses the existing agents to analyze
the audio and video files, and it seems to work well enough for our purposes.
I think we should be able to use the existing agents to analyze
the audio and video files.